Move Windows path logic later in runtime
